Loading the libnanomsg shared library installed into the system
dlopen("/path/libnanomsg.so") is wrong for several reasons.

First, the "*.so" filename is an SDK symlink not present in
production. dlopen() must be prepared to load the versioned
shared library file instead.

Second, the libnanomsg.so embedded into the GIT repo is
an outdated 0.5-beta version of nanomsg with known bugs.

Third, the Mono runtime can be used as embedded Mono where
the native binary part is also linked with nanomsg and
this library instance is obviously the system-preferred
version. In fact, having the binary linked to one version
of a library and dlopen()-ing another different version
leads to symbol conflicts.

To resolve all these issues, NNanomsg needs to support
all different versioned libnanomsg shared libraries and
attempt to open them in order.

To do this properly, dlopen(..., RTLD_NOLOAD) is used first
to ensure that in the embedded Mono case, the library instance
the native binary is linked to is used.

In case the previous step fails, the versioned nanomsg library
files are tried in decreasing order of the version, preferring
the system-installed version first, i.e. the library file name
without the absolute path. The heuristics built into dlopen()
will ensure that on a multilib system, the proper version of
the library is opened.

If the system doesn't have nanomsg installed, the shipped pre-built
binary is loaded as a fallback.

Also support OS-X with a .dylib shared library file name suffix.

Signed-off-by: Böszörményi Zoltán <zboszor@pr.hu>